+++ lib/supply_chain.dart
@@ -6,6 +6,7 @@
 
 library;
 
+export 'src/doc.dart';
 export 'src/graph.dart';
 export 'src/keys.dart';
 export 'src/node_blue_print.dart';
@@ -13,8 +14,8 @@ export 'src/node.dart';
 export 'src/priority.dart';
 export 'src/producer.dart';
 export 'src/scm.dart';
+export 'src/scope_blue_print.dart';
 export 'src/scope.dart';
 export 'src/sub_scope_manager_blue_print.dart';
-export 'src/scope_blue_print.dart';
 export 'src/sub_scope_manager.dart';
 export 'src/tools.dart';